requirements.txt
python-binance
python-dotenv
pandas
pandas-ta
